third_party/blink/web_tests/http/tests/devtools/elements/highlight/highlight-dom-updates-expected.txt

Tests DOM update highlights in the DOM tree.


Running: testDumpInitial
========= Original ========
<No highlights>

Running: testSetAttributeOtherValue
//*[@id="attrTest"]/@attrfoo bar

Running: testSetAttributeEmpty
//*[@id="attrTest"]/@attrfoo (empty)

Running: testAddAttribute
//*[@id="attrTest"]/@attrbar newBar

Running: testRemoveAttribute
//*[@id="attrTest"]

Running: testAppendChildToEmpty
//*[@id="childTest"]

Running: testAppendChildToExpanded
    //*[@id="child2"]

Running: testRemoveChild1
//*[@id="childTest"]

Running: testRemoveChild2
//*[@id="childTest"]

Running: testSetTextContent
//*[@id="textTest"]
//*[@id="textTest"]/text() "Text"

Running: testSetTextNodeTextContent
//*[@id="textTest"]
//*[@id="textTest"]/text() "NewText"

Running: testRemoveInlineTextNode
//*[@id="textTest"]

Running: testSetTextContentWithEmptyText
//*[@id="textTest"]
//*[@id="textTest"]/text() "Text"

Running: testClearTextNodeTextContent
//*[@id="textTest"]

Running: testAppendChildWhenHidden
<No highlights>
